requirements.txt
backcall==0.1.0
backports.weakref==1.0rc1
bleach==1.5.0
boto==2.49.0
boto3==1.9.50
botocore==1.12.50
bz2file==0.98
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
docutils==0.14
gensim==2.0.0
html5lib==0.9999999
idna==2.7
ipdb==0.11
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
jieba==0.38
jmespath==0.9.3
Markdown==3.0.1
numpy==1.15.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
protobuf==3.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.3.0
pypinyin==0.23.0
python-dateutil==2.7.5
requests==2.20.1
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
smart-open==1.7.1
SQLAlchemy==1.2.15
tensorflow-gpu==1.2.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
